pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-incident-response-env"
version = "1.0.0"
description = "OpenEnv incident response environment for SRE workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "incident_response_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "incident_response_env",
    "incident_response_env.server",
    "incident_response_env.server.scenarios",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.2",
]
